use long for timestamp

+++ hal_types.h
@@ -30,6 +30,6 @@ typedef unsigned short u12_t;
 typedef unsigned short u13_t;
 typedef uint32_t u32_t;
 //typedef unsigned int u32_t;
-typedef unsigned int timestamp_t; // WARNING: Must be an unsigned type to properly handle wrapping (u32 wraps in around 1h11m when expressed in us)
+typedef unsigned long timestamp_t; // WARNING: Must be an unsigned type to properly handle wrapping (u32 wraps in around 1h11m when expressed in us)
 
 #endif /* _HAL_TYPES_H_ */
